requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
Django==4.0.2
django-nested-inline==0.4.4
django-object-actions==3.1.0
future==0.18.2
html5lib==1.1
Pillow==9.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
webencodings==0.5.1
xhtml2pdf==0.2.5
